requirements.txt
Django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django-filter
drf-spectacular 
drf-spectacular-sidecar
python-decouple
python-dotenv